Throw 404 when project not found
